+++ include/libpldm/pdr.h
@@ -9,6 +9,22 @@ extern "C" {
 #include <stddef.h>
 #include <stdint.h>
 
+typedef struct pldm_pdr_record {
+	uint32_t record_handle;
+	uint32_t size;
+	uint8_t *data;
+	struct pldm_pdr_record *next;
+	bool is_remote;
+	uint16_t terminus_handle;
+} pldm_pdr_record;
+
+typedef struct pldm_pdr {
+	uint32_t record_count;
+	uint32_t size;
+	pldm_pdr_record *first;
+	pldm_pdr_record *last;
+} pldm_pdr;
+
 /** @struct pldm_pdr
  *  opaque structure that acts as a handle to a PDR repository
  */